Implemented shaders

+++ SimplexCore/SimplexCamera.cs
@@ -28,7 +28,7 @@ namespace SimplexCore
             Position.Y = newY;
             Zoom = newZoom;
 
-            Camera.Position = new Vector2((int)newX, (int)newY);
+            Camera.Position = new Vector2(newX, newY);
             Camera.Zoom = (float)Sgml.clamp(Zoom, 0.05, 30);
             Camera.Rotation = Rotation;
         }
